% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{calc_density}
\alias{calc_density}
\title{Calculate density from movement matrix}
\usage{
calc_density(disperse_mat, sigma = 1e-16)
}
\arguments{
\item{disperse_mat}{Movement matrix}

\item{sigma}{Shift-invert parameter for RSpectra eigs function}
}
\value{
A vector of long-term occupancy densities
}
\description{
Calculate long-term occupancy density from a dispersal matrix
}
\examples{
TBD
}
