% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{calc_step}
\alias{calc_step}
\title{Toy random walk step function}
\usage{
calc_step(
  dist,
  habitat_from,
  habitat_to,
  step_length,
  step_speed,
  pref_strength
)
}
\arguments{
\item{dist}{Distance matrix. A matrix whose entries are every pairwise combination of Euclidean distances}

\item{habitat_from}{type of habitat the step is starting from}

\item{habitat_to}{type of habitat the step is going to}

\item{step_length}{size of step}

\item{step_speed}{speed of step}

\item{pref_strength}{preference strength for habitat types}
}
\value{
A vector of movement probabilities
}
\description{
Performs random walk steps on landscape according to \code{movemodel}
}
\examples{
Used internally by `movemodel`
}
