% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{calculate_clusters}
\alias{calculate_clusters}
\title{Calculate clusters on a landscape}
\usage{
calculate_clusters(
  cluster_type = c("hclust", "DBSCAN", "OPTICS"),
  landscape,
  out,
  min_dens = 1/nrow(landscape),
  n_clust = n_clust,
  ...
)
}
\arguments{
\item{cluster_type}{Specify clustering algorithm, options include "hclust", "DBSCAN", and "OPTICS".
Note: OPTICS has a multi-step workflow which requires user input for each step.}

\item{landscape}{A landscape dataframe with coordinates of points on a landscape.}

\item{out}{A list with five elements:
dists (matrix of kinetic distances as dist object),
occupancy_prob (vector of long term occupancy probabilities for each landscape point),
eps_threshold (lower limit for epsilon parameter if using DBSCAN clustering),
d (number of dimensions used),
T (time scale used)}

\item{min_dens}{Minimum density for a landscape point to be considered for clustering}

\item{n_clust}{Number of clusters if using hclust clustering type}

\item{...}{TBD}
}
\value{
A dataframe with pixel coordinates and cluster assignments
}
\description{
Calculate clusters of organisms on a landscape image based on a movement model and relevant covariates.
Adds a column to the \code{landscape} data.frame specifying which cluster each pixel belongs to.
}
\examples{
Used internally in eigenmove workflow
}
