% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{calculate_eigenfunctions}
\alias{calculate_eigenfunctions}
\title{Calculate eigenvalues and eigenvectors of a movement matrix}
\usage{
calculate_eigenfunctions(movement_matrix, d, discrete_time = FALSE)
}
\arguments{
\item{movement_matrix}{Matrix giving the probability of movement from any point of the landscape to any other point}

\item{d}{Number of dimensions (eigenvalues and eigenvectors) to retain}
}
\value{
A list of eigenvalues and left and right eigenvectors.
The list has four elements:
Phi (matrix of \eqn{d+1} right eigenvectors), Psi (matrix of \eqn{d+1} left eigenvectors),
Lambda (vector of \eqn{d+1} eigenvalues), and d (number of eigenvalues/eigenvectors)
}
\description{
Calculate \eqn{d+1} eigenvalues and left and right eigenvectors
}
\examples{
Used internally in calculate_kinetic_distances()
}
