% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{calculate_kinetic_distances}
\alias{calculate_kinetic_distances}
\title{Calculate kinetic distances}
\usage{
calculate_kinetic_distances(
  movement_matrix,
  d,
  T,
  discrete_time = FALSE,
  scale_by_density = FALSE,
  keep_imaginary = FALSE,
  progress_bar = FALSE
)
}
\arguments{
\item{movement_matrix}{Matrix giving the probability of movement
from any point of the landscape to any other point. A square matrix
whose size is the number of points in the landscape.}

\item{d}{Number of dimensions (eigenvectors/values) to retain}

\item{discrete_time}{Set to TRUE if movement model uses discrete time.}

\item{scale_by_density}{Set to TRUE to scale by the inverse of patch-specific long-term occupancy.}

\item{keep_imaginary}{Set to TRUE to retain complex values in calculation.}

\item{progress_bar}{Set to TRUE to view calculation progress.}
}
\value{
A list with five elements:
dists (matrix of kinetic distances as dist object),
occupancy_prob (vector of long term occupancy probabilities for each landscape point),
eps_threshold (lower limit for epsilon parameter if using DBSCAN clustering),
d (number of dimensions used),
T (time scale used)
}
\description{
Uses the method of Noé et al. (2015) to calculate kinetic distances - lower dimensional distance between two points - based on the eigendecomposition of a movement matrix.
}
\examples{
Used internally in calculate_clusters()
}
