% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{eigenmove}
\alias{eigenmove}
\title{eigenmove}
\description{
Description of your package
}
\details{
_PACKAGE
}
\author{
Spencer Schurman \href{mailto:spencersmtl@gmail.com}{spencersmtl@gmail.com}

Eric Pedersen \href{mailto:eric.pedersen@concordia.ca}{eric.pedersen@concordia.ca}
}
