% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{em_buildgenerator}
\alias{em_buildgenerator}
\title{Build a generator matrix from a movement model and a landscape}
\usage{
em_buildgenerator(landscape, movemodel)
}
\arguments{
\item{landscape}{data frame of landscape pixel coordinates and habitat quality types}

\item{movemodel}{a movement model function that describes individual
movement on a landscape}
}
\value{
A symmetrical CTMC generator matrix: a square matrix giving the
probability (or rate) of movement from any point of the landscape to any
other point. Size is the number of points in the landscape.
}
\description{
Applies a \code{movemodel} function (e.g. produced by \code{em_randomwalk()}) to a
specific landscape to produce a generator matrix.
}
\examples{
movemodel <- em_randomwalk()
# generator <- em_buildgenerator(landscape, movemodel)
}
