% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate-examples.R
\name{em_load_landscape}
\alias{em_load_landscape}
\title{Import a landscape image and convert it to a data frame}
\usage{
em_load_landscape(file, scale = 1, keep_channels = FALSE)
}
\arguments{
\item{file}{Path to an image file. Must be one of a png, jpeg, or bitmap file}

\item{scale}{Scale the image up (>1) or down (<1)}

\item{keep_channels}{Logical value to check if you want to keep all colour channel data, or just greyscale intensities}
}
\value{
A dataframe specifying the coordinates (x and y), colour channel values, greyscale intensity, and habitat quality of each pixel
}
\description{
Converts image file into a dataframe according to pixel coordinates and mean RGBO pixel values. Applies a simple "low", "mid", "high" habitat quality to each pixel depending on mean RGBO value.
}
\examples{
# See vignettes/generate-landscape-vignette.Rmd for example usage
}
