% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate-examples.R
\name{em_neighbourdist}
\alias{em_neighbourdist}
\title{Generate a sparse matrix of nearest neighbour distances}
\usage{
em_neighbourdist(locations, maxdist, nn = 100, ncores = 1)
}
\arguments{
\item{locations}{An sf or sfc object with point geometries}

\item{maxdist}{Maximum distance to consider for neighbours}

\item{nn}{Number of nearest neighbours to consider}

\item{ncores}{Number of cores to use for parallel processing}
}
\value{
A sparse matrix of nearest neighbour distances
}
\description{
This function uses the st_nn function from the nngeo package to find the nearest neighbours of each point within a maximum distance, and returns a sparse matrix of the distances to those neighbours.
}
\examples{
See vignettes/generate-landscape-vignette.Rmd for example usage
}
