% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{em_randomwalk}
\alias{em_randomwalk}
\title{Build a random walk movement model}
\usage{
em_randomwalk(
  step_length = c(0.5, 1, 2),
  step_speed = c(0.5, 1, 2),
  pref_strength = c(4, 2, 1)
)
}
\arguments{
\item{step_length}{vector of step lengths for each habitat quality type}

\item{step_speed}{vector of step speeds for each habitat quality type}

\item{pref_strength}{vector of preference strengths for each habitat quality type}
}
\value{
A function (\code{movemodel}) that,when applied to a \code{landscape} via
\code{em_buildgenerator} returns a generator matrix describing movement rates on
that \code{landscape}
}
\description{
Constructs a movement model function based on a simple random walk model.
Step length, step speed, and habitat preference strength can be adjusted
according to habitat quality type. This functional returns a \code{movemodel}
function that can later be applied to any landscape (e.g. via
\code{em_buildgenerator()}) to produce a generator matrix.
}
\examples{
movemodel <- em_randomwalk()
}
