% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generator.R
\name{Generator}
\alias{Generator}
\title{Generator Class}
\description{
Continuous-time Markov chain describing rates of transition between points on a landscape.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{Generator identifier}

\item{\code{landscape}}{Reference to a Landscape object}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Generator-new}{\code{Generator$new()}}
\item \href{#method-Generator-clone}{\code{Generator$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Generator-new"></a>}}
\if{latex}{\out{\hypertarget{method-Generator-new}{}}}
\subsection{Method \code{new()}}{
Create a new generator
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Generator$new(id, landscape)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{Generator identifier}

\item{\code{landscape}}{A Landscape object}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Generator-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Generator-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Generator$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
