% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{image_to_dataframe}
\alias{image_to_dataframe}
\title{Title}
\usage{
image_to_dataframe(png, scale = 1)
}
\arguments{
\item{png}{An image file}

\item{scale}{Scale the image up (>1) or down (<1)}
}
\value{
A dataframe specifying the habitat quality of each pixel
}
\description{
Title
}
\examples{
insert example here
}
