% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landscape.R
\name{Landscape}
\alias{Landscape}
\title{Landscape Class}
\description{
Object type representing a landscape image in eigenmove
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{The landscape name}

\item{\code{terrain}}{The terrain type}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Landscape-new}{\code{Landscape$new()}}
\item \href{#method-Landscape-clone}{\code{Landscape$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Landscape-new"></a>}}
\if{latex}{\out{\hypertarget{method-Landscape-new}{}}}
\subsection{Method \code{new()}}{
Create a new landscape
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Landscape$new(name, terrain)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Landscape name}

\item{\code{terrain}}{Terrain type}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Landscape-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Landscape-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Landscape$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
