% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate-examples.R
\name{rescale_landscape}
\alias{rescale_landscape}
\title{Rescale continuous-valued landscape to discrete habitat quality types}
\usage{
rescale_landscape(value, good_hab_min = 1, mid_hab_min = 0.5)
}
\arguments{
\item{value}{Continuous-valued landscape variable}

\item{good_hab_min}{Threshold value for high quality habitat}

\item{mid_hab_min}{Threshold value for medium quality habitat}
}
\value{
A factor vector with levels "low", "mid", and "high"
}
\description{
This function re-scales a continuous-valued landscape with a continuous set of values to low, medium, and high values consistent with what is used for the movement model (\code{em_create_example_Q()}), using the case_when function from the dplyr package.
}
\examples{
See vignettes/generate-landscape-vignette.Rmd for example usage
}
