% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate-examples.R
\name{sparse_dispersemat}
\alias{sparse_dispersemat}
\title{Generate a sparse dispersal matrix}
\usage{
sparse_dispersemat(
  nn_distmat,
  patch_qual,
  d0,
  qual_bias,
  dist_effect,
  alpha,
  lambda,
  qual0,
  dmax,
  dmin = 1e-12
)
}
\arguments{
\item{nn_distmat}{Sparse matrix of nearest neighbour distances}

\item{patch_qual}{Vector of patch qualities}

\item{d0}{Base dispersal rate}

\item{qual_bias}{Quality bias parameter}

\item{dist_effect}{Distance effect parameter}

\item{alpha}{Parameter used in calculation of base movement rate}

\item{lambda}{Parameter used in calculation of base movement rate}

\item{qual0}{Parameter used in calculation of base movement rate}

\item{dmax}{Maximum movement rate}

\item{dmin}{Minimum movement rate}
}
\value{
A sparse dispersal matrix
}
\description{
This function generates a sparse dispersal matrix based on a nearest neighbour distance matrix and patch qualities.
}
\examples{
TBD
}
