% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{check_conjugate_state}
\alias{check_conjugate_state}
\title{Check conjugate state of eigenvalues}
\usage{
check_conjugate_state(eigenvalues, tol = 1e-14)
}
\arguments{
\item{eigenvalues}{A vector of eigenvalues}

\item{tol}{tolerance defining which numbers are considered zero}
}
\value{
Binary value indicating whether there are complex eigenvalues and if they are organized correctly
}
\description{
This function checks that if the smallest eigenvalue has a non-zero imaginary component, that imaginary component is matched with a conjugate eigenvalue in the next smallest value. Otherwise, the kinetic distance calculation won't result in all real kinetic distances.
}
\examples{
Used internally in calculate_kinetic_distances()
}
