% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate-examples.R
\name{create_GP_landscape}
\alias{create_GP_landscape}
\title{Simulate a toy landscape using a Gaussian process}
\usage{
create_GP_landscape(
  landscape_width = 10,
  landscape_height = 10,
  patch_scale = 1
)
}
\arguments{
\item{landscape_width}{width of image in pixels}

\item{landscape_height}{height of image in pixels}

\item{patch_scale}{size of the patch. Larger values of patch_scale correspond to higher correlations between distant points, so larger (and fewer) patches}
}
\value{
A dataframe specifying the coordinates (x and y) and intensity value of each pixel
}
\description{
Simulate GP landscapes for use with simple random walk toy movement models. A Gaussian process  assumes that a random value (habitat quality) at each point in a landscape follows a normal distribution, and that the random values for points close to one another are correlated, so they have similar values.
}
\examples{
See vignettes/generate-landscape-vignette.Rmd for example usage
}
