% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate-examples.R
\name{em_create_example_Q}
\alias{em_create_example_Q}
\title{Generate a movement matrix given a landscape image input and step parameters}
\usage{
em_create_example_Q(
  landscape,
  step_length = c(0.5, 0.5, 2),
  speed = c(0.5, 0.5, 2),
  pref_strength = c(4, 2, 1)
)
}
\arguments{
\item{landscape}{dataframe of landscape pixel coordinates and habitat quality types}

\item{step_length}{vector of step lengths for each habitat quality type}

\item{speed}{vector of step speeds for each habitat quality type}

\item{pref_strength}{vector of preference strengths for each habitat quality type}
}
\value{
A symmetrical movement matrix: A square matrix giving the probability of movement from any point of the landscape to any other point. Size is the number of points in the landscape.
}
\description{
Generates a movement matrix based on a simple random walk model where step length, speed, and preference strength can be adjusted according to habitat quality types defined in the landscape dataframe.
}
\examples{
See vignettes/generate-landscape-vignette.Rmd for example usage
}
