% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate-examples.R
\name{em_load_landscape}
\alias{em_load_landscape}
\title{Import an image file of an example landscape and convert it to a data frame}
\usage{
em_load_landscape(file, scale = 1, keep_channels = FALSE)
}
\arguments{
\item{file}{Path to an image file. Must be one of a png, jpeg, or bitmap file}

\item{scale}{Scale the image up (>1) or down (<1)}
}
\value{
A dataframe specifying the coordinates (x and y), colour channel values, greyscale intensity, and habitat quality of each pixel
}
\description{
Import an image file of an example landscape and convert it to a data frame
}
\examples{
# insert example
}
