% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{em_loadlandscape}
\alias{em_loadlandscape}
\title{Import a landscape as a data frame}
\usage{
em_loadlandscape(file, scale = 1, keep_channels = FALSE)
}
\arguments{
\item{file}{Path to an image file. Must be one of a png, jpeg, or bitmap file}

\item{scale}{Scale the image up (>1) or down (<1)}

\item{keep_channels}{Logical value to check if you want to keep all colour channel data, or just greyscale intensities}
}
\value{
A data frame specifying the coordinates (x and y) and pixel values
}
\description{
Converts image file into a data frame with columns for x and y pixel cartesian coordinates, greyscale intensities and optionally RGBO color channels
}
\examples{
# See vignettes/generate-landscape-vignette.Rmd or eigenmove-vignette.Rmd for example usage
}
